Create README.md
In this dashboard, there is an interactive dashboard which contains the visualization of the different beverage companies' revenue generated from the sales and the units sold. The overall distribution in sales across all of the states within the United States are shown with bubbles and there is a zoom in and out button for viewing this.